[FSA] Add ID feature to file picker
Allows for specification of the "purpose" of a file picker invocation.
When an ID is specified, the picker will remember the picked directory.
The next time the matching ID is specified, the picker will default to
this directory.

This ID is remembered on a per-origin basis.

Specifying an empty ID (id: "") is the same as not specifying an ID.

Explainer:
https://github.com/WICG/file-system-access/blob/main/SuggestedNameAndDir.md#distinguishing-the-purpose-of-different-file-picker-invocations

Bug: 1173164
Change-Id: I89fce44efd3eb425a0ea8d10192b71758933052d